python/requirement.txt
langchain
mysql-connector-python
langchain_community
langchain_chroma
langchain_openai